refactor: folder structure of examples/web
Modify folder names

Closes #21
