+++ examples/web/src/common_components.rs
