+++ examples/web/src/common_components/Main.rs
@@ -1,13 +1,13 @@
 #![allow(non_snake_case)]
 use dioxus::prelude::*;
 
-use crate::pages;
+use crate::pages::home::HomePage;
 
 pub fn Main(cx: Scope) -> Element {
     cx.render(rsx! {
             main {
             class: "w-full relative",
-            pages::Home::Home {}
+            HomePage::HomePage {}
         }
     })
 }
