+++ examples/web/src/common_components/Nav.rs
