+++ examples/web/src/pages.rs
@@ -1,2 +1,3 @@
 #![allow(non_snake_case)]
-pub mod Home;
+
+pub mod home;
